requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.30.1
pydantic==2.7.4
orjson==3.10.15
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9
PyYAML==6.0.2
torch==2.5.1
transformers==4.55.0
gliner2
spacy==3.7.5
python-docx==1.1.2
huggingface_hub==0.35.0
filelock==3.16.1
nltk==3.9.1
pymupdf==1.25.1
